# mfdc-nc/docs/swagger.yaml
definitions:
  model.ActiveTeamNumber:
    properties:
      activated_at:
        type: string
      expired_at:
        type: string
      expired_at_unixtime:
        type: integer
      name:
        type: string
      number:
        type: string
      spin:
        type: integer
      team_id:
        type: integer
      vendor_name:
        type: string
    type: object
  model.DayLogs:
    properties:
      end_at:
        type: string
      start_at:
        type: string
    type: object
  model.DayNumbers:
    properties:
      active:
        type: boolean
      enabled:
        type: boolean
      id:
        type: integer
      last_activated_at:
        type: string
      logs:
        items:
          $ref: '#/definitions/model.DayLogs'
        type: array
      marked:
        type: boolean
      number:
        type: string
      spin:
        type: integer
      used:
        type: boolean
    type: object
  model.DayPools:
    properties:
      active:
        type: boolean
      active_subpool:
        $ref: '#/definitions/model.DaySubPool'
      block:
        type: integer
      created_at:
        type: string
      finish:
        type: boolean
      finish_at:
        type: string
      id:
        type: integer
      name:
        type: string
      num_count:
        type: integer
      rotation:
        type: boolean
      vendor_id:
        type: integer
    type: object
  model.DaySubPool:
    properties:
      activated_at:
        type: string
      id:
        type: integer
      numbers:
        items:
          $ref: '#/definitions/model.DayNumbers'
        type: array
      spin:
        type: integer
    type: object
...
  model.Reload:
    properties:
      reload:
        type: string
    type: object
  model.Resource:
    properties:
      resources:
        items:
          type: integer
        type: array
      vendor_id:
        type: integer
    type: object
...
  model.SwaggerActiveTeamNumber:
    properties:
      data:
        items:
          $ref: '#/definitions/model.ActiveTeamNumber'
        type: array
      status:
        type: string
    type: object
  model.SwaggerDefaultResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  model.SwaggerTeamDayNumbers:
    properties:
      data:
        items:
          $ref: '#/definitions/model.TeamDayNumbers'
        type: array
      status:
        type: string
    type: object
  model.SwaggerTeams:
    properties:
      actual_vendor_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      webitel_res_ids:
        $ref: '#/definitions/model.Resource'
    type: object
  model.SwaggerTeamsList:
    properties:
      data:
        items:
          $ref: '#/definitions/model.SwaggerTeams'
        type: array
      status:
        type: string
    type: object
...
  model.TeamDayNumbers:
    properties:
      name:
        type: string
      pools:
        items:
          $ref: '#/definitions/model.DayPools'
        type: array
      team_id:
        type: integer
    type: object
  model.TeamsDayRequest:
    properties:
      from_date:
        type: string
      to_date:
        type: string
    type: object
...
info:
  contact: {}
  description: Swagger API for Golang Project MFDC NC
  title: MFDC Number changer
  version: "1.0"
paths:
  /config/reload:
    get:
      consumes:
      - application/json
      description: Reloading API configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Reload'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Execute reload config
      tags:
      - Reload
...
  /teams/{id}/rotate:
    get:
      consumes:
      - application/json
      description: Rotate team number
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerDefaultResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Rotate team number
      tags:
      - Teams
  /teams/activenums:
    get:
      consumes:
      - application/json
      description: Active teams numbers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerActiveTeamNumber'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Active teams numbers
      tags:
      - Teams
  /teams/add:
    post:
      consumes:
      - application/json
      description: |-
        Add new team
        JSON object containing resource IDs
      parameters:
      - description: Team
        in: body
        name: route
        required: true
        schema:
          $ref: '#/definitions/model.SwaggerTeams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerDefaultResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Add team
      tags:
      - Teams
  /teams/daynums:
    post:
      consumes:
      - application/json
      description: List teams numbers per day
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.TeamsDayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerTeamDayNumbers'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List teams numbers per day
      tags:
      - Teams
  /teams/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Drop team
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerDefaultResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Delete team
      tags:
      - Teams
  /teams/edit/{id}:
    patch:
      consumes:
      - application/json
      description: |-
        Edit team params
        JSON object containing resource IDs
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data without ID
        in: body
        name: ip
        required: true
        schema:
          $ref: '#/definitions/model.SwaggerTeams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerDefaultResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Edt team
      tags:
      - Teams
  /teams/list:
    get:
      consumes:
      - application/json
      description: List teams
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerTeamsList'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List teams
      tags:
      - Teams
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-MFDC-Key
    type: apiKey
swagger: "2.0"